# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "PixelGhost API", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Read Root", "description": "Health check endpoint.", "operationId": "read_root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/tasks": {"get": {"summary": "List All Tasks", "description": "List all submitted tasks with their IDs.", "operationId": "list_all_tasks_tasks_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/task/status/{task_id}": {"get": {"summary": "Get Task Status", "description": "Get the status of a submitted task.", "operationId": "get_task_status_task_status__task_id__get", "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/task/result/{task_id}": {"get": {"summary": "Get Task Result", "description": "Get the result of a completed task (e.g., download the output file).", "operationId": "get_task_result_task_result__task_id__get", "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/text/lsb/encode": {"post": {"summary": "Encode Text In Image", "description": "Encode a hidden text message into an image using LSB steganography.", "operationId": "encode_text_in_image_text_lsb_encode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_encode_text_in_image_text_lsb_encode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/text/lsb/decode": {"post": {"summary": "Decode Text From Image", "description": "Decode and extract a hidden text message from an encoded image.", "operationId": "decode_text_from_image_text_lsb_decode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_decode_text_from_image_text_lsb_decode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/text/lsb_random/encode": {"post": {"summary": "Lsb Random Encode Text In Image", "description": "Encode a hidden text message into an image using LSB steganography with randomized pixel selection.", "operationId": "lsb_random_encode_text_in_image_text_lsb_random_encode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_lsb_random_encode_text_in_image_text_lsb_random_encode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/text/lsb_random/decode": {"post": {"summary": "Lsb Random Decode Text From Image", "description": "Decode and extract a hidden text message from an encoded image using LSB steganography with randomized pixel selection.", "operationId": "lsb_random_decode_text_from_image_text_lsb_random_decode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_lsb_random_decode_text_from_image_text_lsb_random_decode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/text/lsb_random_enc/encode": {"post": {"summary": "Lsb Random Enc Encode Text In Image", "description": "Encode and encrypt a hidden text message into an image using LSB steganography with randomized pixel selection.", "operationId": "lsb_random_enc_encode_text_in_image_text_lsb_random_enc_encode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_lsb_random_enc_encode_text_in_image_text_lsb_random_enc_encode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/text/lsb_random_enc/decode": {"post": {"summary": "Lsb Random Enc Decode Text From Image", "description": "Decode and decrypt a hidden text message from an encoded image using LSB steganography with randomized pixel selection.", "operationId": "lsb_random_enc_decode_text_from_image_text_lsb_random_enc_decode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_lsb_random_enc_decode_text_from_image_text_lsb_random_enc_decode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/image/lsb/encode": {"post": {"summary": "Encode Image In Image", "description": "Encode a secret image into a cover image using basic LSB steganography.", "operationId": "encode_image_in_image_image_lsb_encode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_encode_image_in_image_image_lsb_encode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/image/lsb/decode": {"post": {"summary": "Decode Image From Image", "description": "Decode and extract a hidden image from a stego image using basic LSB steganography.", "operationId": "decode_image_from_image_image_lsb_decode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_decode_image_from_image_image_lsb_decode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/image/lsb_random/encode": {"post": {"summary": "Encode Image In Image Random", "description": "Encode a secret image into a cover image using LSB steganography with pseudorandom pixel selection.", "operationId": "encode_image_in_image_random_image_lsb_random_encode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_encode_image_in_image_random_image_lsb_random_encode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/image/lsb_random/decode": {"post": {"summary": "Decode Image From Image Random", "description": "Decode and extract a hidden image from a stego image using LSB steganography with pseudorandom pixel selection.", "operationId": "decode_image_from_image_random_image_lsb_random_decode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_decode_image_from_image_random_image_lsb_random_decode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/image/lsb_random_enc/encode": {"post": {"summary": "Encode Image In Image Encrypted", "description": "Encode and encrypt a secret image into a cover image using LSB steganography with pseudorandom pixel selection.", "operationId": "encode_image_in_image_encrypted_image_lsb_random_enc_encode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_encode_image_in_image_encrypted_image_lsb_random_enc_encode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/image/lsb_random_enc/decode": {"post": {"summary": "Decode Image From Image Encrypted", "description": "Decode and decrypt a hidden image from a stego image using LSB steganography with pseudorandom pixel selection.", "operationId": "decode_image_from_image_encrypted_image_lsb_random_enc_decode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_decode_image_from_image_encrypted_image_lsb_random_enc_decode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/image/dct/encode": {"post": {"summary": "Dct Encode Image", "description": "Encode a secret image into a cover image using DCT steganography.", "operationId": "dct_encode_image_image_dct_encode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_dct_encode_image_image_dct_encode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/image/dct/decode": {"post": {"summary": "Dct Decode Image", "description": "Decode and extract a hidden image from an encoded image using DCT steganography.", "operationId": "dct_decode_image_image_dct_decode_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_dct_decode_image_image_dct_decode_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_dct_decode_image_image_dct_decode_post": {"properties": {"image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Image"}}, "type": "object", "required": ["image"], "title": "Body_dct_decode_image_image_dct_decode_post"}, "Body_dct_encode_image_image_dct_encode_post": {"properties": {"cover_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Cover Image"}, "secret_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Secret Image"}}, "type": "object", "required": ["cover_image", "secret_image"], "title": "Body_dct_encode_image_image_dct_encode_post"}, "Body_decode_image_from_image_encrypted_image_lsb_random_enc_decode_post": {"properties": {"stego_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Stego Image"}, "key": {"type": "string", "title": "Key"}}, "type": "object", "required": ["stego_image", "key"], "title": "Body_decode_image_from_image_encrypted_image_lsb_random_enc_decode_post"}, "Body_decode_image_from_image_image_lsb_decode_post": {"properties": {"stego_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Stego Image"}}, "type": "object", "required": ["stego_image"], "title": "Body_decode_image_from_image_image_lsb_decode_post"}, "Body_decode_image_from_image_random_image_lsb_random_decode_post": {"properties": {"stego_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Stego Image"}, "key": {"type": "string", "title": "Key"}}, "type": "object", "required": ["stego_image", "key"], "title": "Body_decode_image_from_image_random_image_lsb_random_decode_post"}, "Body_decode_text_from_image_text_lsb_decode_post": {"properties": {"image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Image"}}, "type": "object", "required": ["image"], "title": "Body_decode_text_from_image_text_lsb_decode_post"}, "Body_encode_image_in_image_encrypted_image_lsb_random_enc_encode_post": {"properties": {"cover_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Cover Image"}, "secret_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Secret Image"}, "key": {"type": "string", "title": "Key"}}, "type": "object", "required": ["cover_image", "secret_image", "key"], "title": "Body_encode_image_in_image_encrypted_image_lsb_random_enc_encode_post"}, "Body_encode_image_in_image_image_lsb_encode_post": {"properties": {"cover_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Cover Image"}, "secret_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Secret Image"}}, "type": "object", "required": ["cover_image", "secret_image"], "title": "Body_encode_image_in_image_image_lsb_encode_post"}, "Body_encode_image_in_image_random_image_lsb_random_encode_post": {"properties": {"cover_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Cover Image"}, "secret_image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Secret Image"}, "key": {"type": "string", "title": "Key"}}, "type": "object", "required": ["cover_image", "secret_image", "key"], "title": "Body_encode_image_in_image_random_image_lsb_random_encode_post"}, "Body_encode_text_in_image_text_lsb_encode_post": {"properties": {"image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Image"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["image", "message"], "title": "Body_encode_text_in_image_text_lsb_encode_post"}, "Body_lsb_random_decode_text_from_image_text_lsb_random_decode_post": {"properties": {"image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Image"}, "key": {"type": "string", "title": "Key"}}, "type": "object", "required": ["image", "key"], "title": "Body_lsb_random_decode_text_from_image_text_lsb_random_decode_post"}, "Body_lsb_random_enc_decode_text_from_image_text_lsb_random_enc_decode_post": {"properties": {"image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Image"}, "key": {"type": "string", "title": "Key"}}, "type": "object", "required": ["image", "key"], "title": "Body_lsb_random_enc_decode_text_from_image_text_lsb_random_enc_decode_post"}, "Body_lsb_random_enc_encode_text_in_image_text_lsb_random_enc_encode_post": {"properties": {"image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Image"}, "message": {"type": "string", "title": "Message"}, "key": {"type": "string", "title": "Key"}}, "type": "object", "required": ["image", "message", "key"], "title": "Body_lsb_random_enc_encode_text_in_image_text_lsb_random_enc_encode_post"}, "Body_lsb_random_encode_text_in_image_text_lsb_random_encode_post": {"properties": {"image": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Image"}, "message": {"type": "string", "title": "Message"}, "key": {"type": "string", "title": "Key"}}, "type": "object", "required": ["image", "message", "key"], "title": "Body_lsb_random_encode_text_in_image_text_lsb_random_encode_post"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}